Refactor into types
