+++ crypto/src/types/mod.rs
@@ -1,0 +1,101 @@
+mod contribution;
+mod error;
+mod transcript;
+mod utils;
+
+use crate::{
+    crypto::g1_mul_glv, g1_subgroup_check, g2_subgroup_check, parse_g, zcash_format::write_g,
+    ParseError,
+};
+use ark_bls12_381::{g2, Bls12_381, Fr, G1Affine, G1Projective, G2Affine, G2Projective};
+use ark_ec::{
+    msm::VariableBaseMSM, short_weierstrass_jacobian::GroupAffine, AffineCurve, PairingEngine,
+    ProjectiveCurve, SWModelParameters,
+};
+use ark_ff::{One, PrimeField, UniformRand, Zero};
+use rayon::prelude::*;
+use serde::{Deserialize, Serialize};
+use std::{cmp::max, iter};
+use thiserror::Error;
+use tracing::{error, instrument};
+use zeroize::Zeroizing;
+
+pub use self::{
+    contribution::{Contribution, SubContribution},
+    error::{CeremoniesError, CeremonyError},
+    transcript::{SubTranscript, Transcript},
+};
+
+#[cfg(test)]
+pub mod test {
+    use ark_ff::UniformRand;
+
+    use super::*;
+
+    #[test]
+    fn verify() {
+        let mut transcript = SubTranscript::new(32768, 65);
+        let mut contrib = SubContribution::new(32768, 65);
+        assert!(contrib.verify(&transcript));
+        let mut rng = rand::thread_rng();
+        contrib.add_tau(&Fr::rand(&mut rng));
+        assert!(contrib.verify(&transcript));
+    }
+}
+
+#[cfg(feature = "bench")]
+#[doc(hidden)]
+pub mod bench {
+    use ark_ff::UniformRand;
+    use criterion::{black_box, BatchSize, BenchmarkId, Criterion};
+
+    use crate::bench::rand_fr;
+
+    use super::*;
+
+    pub fn group(criterion: &mut Criterion) {
+        bench_pow_tau(criterion);
+        bench_add_tau(criterion);
+        bench_verify(criterion);
+    }
+
+    fn bench_pow_tau(criterion: &mut Criterion) {
+        criterion.bench_function("contribution/pow_tau", move |bencher| {
+            let mut rng = rand::thread_rng();
+            let tau = Zeroizing::new(Fr::rand(&mut rng));
+            bencher.iter(|| black_box(SubContribution::pow_table(black_box(&tau), 32768)));
+        });
+    }
+
+    fn bench_add_tau(criterion: &mut Criterion) {
+        for size in crate::SIZES {
+            criterion.bench_with_input(
+                BenchmarkId::new("contribution/add_tau", format!("{:?}", size)),
+                &size,
+                move |bencher, (n1, n2)| {
+                    let mut contrib = SubContribution::new(*n1, *n2);
+                    bencher.iter_batched(
+                        || rand_fr(),
+                        |tau| contrib.add_tau(&tau),
+                        BatchSize::SmallInput,
+                    );
+                },
+            );
+        }
+    }
+
+    fn bench_verify(criterion: &mut Criterion) {
+        for size in crate::SIZES {
+            criterion.bench_with_input(
+                BenchmarkId::new("contribution/verify", format!("{:?}", size)),
+                &size,
+                move |bencher, (n1, n2)| {
+                    let mut transcript = SubTranscript::new(*n1, *n2);
+                    let mut contrib = SubContribution::new(*n1, *n2);
+                    contrib.add_tau(&rand_fr());
+                    bencher.iter(|| black_box(contrib.verify(&transcript)));
+                },
+            );
+        }
+    }
+}
